+++ packages/url-loader/src/index.ts
@@ -4,7 +4,7 @@ export type { ConstructUrlProps, PluginOptionsResize, PluginOptions, PluginResul
 export { effects, position, primary, text } from './constants/qualifiers';
 
 export type { AssetOptionsResize, AssetOptions } from './types/asset';
-export type { ImageOptionsResize, ImageOptionsZoomPan, ImageOptions } from './types/image';
+export type { ImageOptionsFillBackground, ImageOptionsResize, ImageOptionsZoomPan, ImageOptions } from './types/image';
 export type { VideoOptionsResize, VideoOptions } from './types/video';
 
 export type { AnalyticsOptions } from './types/analytics';