+++ packages/url-loader/src/types/image.ts
@@ -2,6 +2,14 @@ import { z } from 'zod';
 
 import { assetOptionsSchema } from './asset';
 
+export const imageOptionsFillBackgroundSchema = z.object({
+  crop: z.string().optional(),
+  gravity: z.string().optional(),
+  prompt: z.string().optional()
+})
+
+export type ImageOptionsFillBackground = z.infer<typeof imageOptionsFillBackgroundSchema>;
+
 export const imageOptionsResizeSchema = assetOptionsSchema;
 
 export type ImageOptionsResize = z.infer<typeof imageOptionsResizeSchema>;
@@ -13,7 +21,17 @@ export const imageOptionsZoomPanSchema = z.object({
 
 export type ImageOptionsZoomPan = z.infer<typeof imageOptionsZoomPanSchema>;
 
+// ImageOptionsFillBackground
 export const imageOptionsSchema = assetOptionsSchema.extend({
+  fillBackground: z.union([
+      z.boolean(),
+      imageOptionsFillBackgroundSchema
+    ])
+    .describe(JSON.stringify({
+      text: 'Uses Generative Fill to extended padded image with AI',
+      url: 'https://cloudinary.com/documentation/transformation_reference#b_gen_fill'
+    }))
+    .optional(),
   zoompan: z.union([
       z.string(),
       z.boolean(),