+++ packages/util/src/index.ts
@@ -1,4 +1,6 @@
 export { parseUrl, getPublicId, getTransformations } from './lib/cloudinary';
 export type { ParseUrl } from './lib/cloudinary';
 
+export { testColorIsHex, convertColorHexToRgb } from './lib/colors';
+
 export { encodeBase64, objectHasKey, sortByKey } from './lib/util';