+++ packages/util/src/lib/cloudinary.ts
@@ -1,6 +1,6 @@
 const REGEX_VERSION = /\/v\d+\//;
 const REGEX_FORMAT = /\.(ai|avif|gif|png|webp|bmp|bw|djvu|dng|ps|ept|eps|eps3|fbx|flif|gif|glb|gltf|heif|heic|ico|indd|jpg|jpe|jpeg|jp2|wdp|jxr|hdp|obj|pdf|ply|png|psd|arw|cr2|svg|tga|tif|tiff|u3ma|usdz|webp|3g2|3gp|avi|flv|m3u8|ts|m2ts|mts|mov|mkv|mp4|mpeg|mpd|mxf|ogv|webm|wmv)$/i
-const REGEX_URL = /https?:\/\/(?<host>[^\/]+)\/(?<cloudName>[^\/]+)\/(?<assetType>image|images|video|videos|raw|files)\/(?<deliveryType>upload|fetch|private|authenticated|sprite|facebook|twitter|youtube|vimeo)?\/?(?<signature>s\-\-[a-zA-Z0-9]+\-\-)?\/?(?<transformations>(?:[^_\/]+_[^,\/]+,?\/?)*\/)*(?<version>v\d+|\w{1,2})\/(?<publicId>[^\s]+)$/;
+const REGEX_URL = /https?:\/\/(?<host>[^\/]+)\/(?<cloudName>[^\/]+)\/(?<assetType>image|images|video|videos|raw|files)\/(?<deliveryType>upload|fetch|private|authenticated|sprite|facebook|twitter|youtube|vimeo)?\/?(?<signature>s--([a-zA-Z0-9\_\-]{8}|[a-zA-Z0-9\_\-]{32})--)?\/?(?<transformations>(?:[^_\/]+_[^,\/]+,?\/?)*\/)*(?<version>v\d+|\w{1,2})\/(?<publicId>[^\s]+)$/;
 const ASSET_TYPES_SEO = ['images', 'videos', 'files'];
 
 
@@ -75,6 +75,14 @@ export function parseUrl(src: string): ParseUrl | undefined {
     const publicIdParts = parts.publicId?.split('/') || [];
     parts.seoSuffix = publicIdParts.pop();
     parts.publicId = publicIdParts.join('/');
+  }
+
+  // The URL Gen SDK which this library relies on will re-encode the public ID. To avoid issues where
+  // someone is already passing in a URL or ID that's been encoded programmatically, first decode
+  // the public ID, which should theoretically be harmless since it ends up getting encoded
+
+  if ( parts.publicId ) {
+    parts.publicId = decodeURIComponent(parts.publicId);
   }
 
   return parts;